vision_assistant/requirements.txt
SpeechRecognition
pyaudio
pyttsx3
openai
python-dotenv
colorama
